pyproject.toml
[project]
name = "ara-dialectical-fragments"
version = "0.1.0"
description = ""
authors = [
    {name = "Kirill Chirkunov",email = "kirill.chirkunov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "fasttext-langdetect (>=1.0.5,<2.0.0)",
    "lingua-language-detector (>=2.1.1,<3.0.0)",
    "datasets (>=4.0.0,<5.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
